Basic upload
Includes main menu, law of cosines, heron's formula, and law of sines…
I'm getting the wrong numbers from either all of them, or just a few.
There is an unimplemented bit for law of sines, and the end menus need
to be updated to work with the new main menu system.